+++ BackEnd/CmApp/Controllers/TrackingController.cs
@@ -1,13 +1,16 @@
 ﻿using System;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using CmApp.Entities;
 using CmApp.Repositories;
 using CmApp.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CmApp.Controllers
 {
     [Route("api/cars/{carId}/tracking")]
+    [Authorize(Roles = "user, admin", AuthenticationSchemes = "user, admin")]
     [ApiController]
     public class TrackingController : ControllerBase
     {
@@ -18,44 +21,94 @@ namespace CmApp.Controllers
             ScraperService = new WebScraper(),
             FileRepository = new FileRepository()
         };
+        private readonly CarRepository carRepo = new CarRepository();
 
         // GET: api/cars/{carId}/tracking
         [HttpGet]
-        public async Task<TrackingEntity> Get(string carId)
+        [Authorize(Roles = "user, admin")]
+        public async Task<IActionResult> Get(string carId)
         {
-            var tracking = await trackingService.GetTracking(carId);
-            return tracking;
+            try
+            {
+                var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                var role = HttpContext.User.FindFirst(ClaimTypes.Role).Value;
+                var car = await carRepo.GetCarById(carId);
+                if (car.User != userId && role != "admin")
+                    throw new Exception("Car does not exist");
+
+                var tracking = await trackingService.GetTracking(carId);
+                return Ok(tracking);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         // POST: api/cars/{carId}/tracking
         [HttpPost]
+        [Authorize(Roles = "user, admin")]
         public async Task<IActionResult> Post(string carId)
         {
             try
             {
+                var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                var role = HttpContext.User.FindFirst(ClaimTypes.Role).Value;
+                var car = await carRepo.GetCarById(carId);
+                if (car.User != userId && role != "admin")
+                    throw new Exception("Car does not exist");
+
                 var newTracking = await trackingService.LookForTracking(carId);
-                return StatusCode(200, newTracking);
+                return Ok(newTracking);
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
-            }          
+                return BadRequest(ex.Message);
+            }
         }
 
         // PUT: api/cars/{carId}/tracking/
         [HttpPut]
-        public async Task<NoContentResult> Put(string carId, [FromBody] TrackingEntity tracking)
+        [Authorize(Roles = "user, admin")]
+        public async Task<IActionResult> Put(string carId, [FromBody] TrackingEntity tracking)
         {
-            await trackingService.UpdateTracking(carId, tracking);
-            return NoContent();
+            try
+            {
+                var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                var role = HttpContext.User.FindFirst(ClaimTypes.Role).Value;
+                var car = await carRepo.GetCarById(carId);
+                if (car.User != userId && role != "admin")
+                    throw new Exception("Car does not exist");
+
+                await trackingService.UpdateTracking(carId, tracking);
+                return NoContent();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         // DELETE: api/cars/{carId}/tracking/
         [HttpDelete]
-        public async Task<NoContentResult> Delete(string carId)
+        [Authorize(Roles = "user, admin")]
+        public async Task<IActionResult> Delete(string carId)
         {
-            await trackingService.DeleteTracking(carId);
-            return NoContent();
+            try
+            {
+                var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                var role = HttpContext.User.FindFirst(ClaimTypes.Role).Value;
+                var car = await carRepo.GetCarById(carId);
+                if (car.User != userId && role != "admin")
+                    throw new Exception("Car does not exist");
+
+                await trackingService.DeleteTracking(carId);
+                return NoContent();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
     }
 }
